requirements.txt
arabic-reshaper==2.1.3
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
certifi==2021.5.30
crispy-bootstrap5==0.6
cycler==0.10.0
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
fonttools==4.25.0
future==0.18.2
html5lib==1.1
kiwisolver @ file:///C:/ci/kiwisolver_1612282606037/work
matplotlib @ file:///C:/ci/matplotlib-suite_1628011511716/work
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
munkres==1.1.4
numpy @ file:///C:/ci/numpy_and_numpy_base_1626271903025/work
olefile==0.46
pandas @ file:///C:/ci/pandas_1618365634936/work
Pillow @ file:///C:/ci/pillow_1625663293593/work
psycopg2 @ file:///C:/ci/psycopg2_1612298803839/work
psycopg2-binary==2.9.1
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
reportlab==3.6.1
scipy @ file:///C:/ci/scipy_1618856128765/work
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///C:/ci/tornado_1606942392901/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
webencodings==0.5.1
wincertstore==0.2
xhtml2pdf==0.2.5
